+++ voting-booth-frontend/src/app/services/candidacy/candidacy.service.ts
@@ -3,7 +3,7 @@ import {HttpClient} from '@angular/common/http';
 import {first} from 'rxjs/operators';
 import {Store} from '@ngrx/store';
 import {AppState} from '../../reducers';
-import {LoadCandidateAction} from '../../actions/candidate.actions';
+import {LoadCandidateAction} from '../../actions/candidate/candidate.actions';
 import {CandidateData} from '../../types';
 
 @Injectable({
